# docs/public/api/openapi-3.0.yaml
openapi: 3.0.0
info:
  description: Track, Manage, and Organize your Things.
  title: Homebox API
  contact:
    name: Homebox Team
    url: https://discord.homebox.software
  version: "1.0"
paths:
...
  /v1/templates:
    get:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Get All Entity Templates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.EntityTemplateSummary"
    post:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Create Entity Template
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityTemplateCreate"
        description: Template Data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityTemplateOut"
  "/v1/templates/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Get Entity Template
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityTemplateOut"
    put:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Update Entity Template
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/repo.EntityTemplateUpdate"
        description: Template Data
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityTemplateOut"
    delete:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Delete Entity Template
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
  "/v1/templates/{id}/batch-create":
    post:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Batch Create Entities from Template
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.EntityTemplateBatchCreateRequest"
        description: Batch options
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/repo.EntityOut"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
  "/v1/templates/{id}/create-item":
    post:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Create Entity from Template
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/v1.EntityTemplateCreateItemRequest"
        description: Entity Data
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityOut"
  "/v1/templates/{id}/photo":
    get:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Get Template Photo
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/octet-stream:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
    post:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Upload Template Photo
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                file:
                  description: Photo file
                  type: string
                  format: binary
              required:
                - file
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/repo.EntityTemplateOut"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
    delete:
      security:
        - Bearer: []
      tags:
        - Entity Templates
      summary: Delete Template Photo
      parameters:
        - description: Template ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validate.ErrorResponse"
...
servers:
  - url: /api
components:
  securitySchemes:
    Bearer:
      description: "\"Type 'Bearer TOKEN' to correctly set the API Key\""
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    ent.Attachment:
      type: object
      properties:
        created_at:
          description: CreatedAt holds the value of the "created_at" field.
          type: string
        edges:
          description: >-
            Edges holds the relations/edges for other nodes in the graph.

            The values are being populated by the AttachmentQuery when eager-loading is set.
          allOf:
            - $ref: "#/components/schemas/ent.AttachmentEdges"
        id:
          description: ID of the ent.
          type: string
        mime_type:
          description: MimeType holds the value of the "mime_type" field.
          type: string
        path:
          description: Path holds the value of the "path" field.
          type: string
        primary:
          description: Primary holds the value of the "primary" field.
          type: boolean
        title:
          description: Title holds the value of the "title" field.
          type: string
        type:
          description: Type holds the value of the "type" field.
          allOf:
            - $ref: "#/components/schemas/attachment.Type"
        updated_at:
          description: UpdatedAt holds the value of the "updated_at" field.
          type: string
...
    repo.EntityFieldData:
      type: object
      properties:
        booleanValue:
          type: boolean
        id:
          type: string
        name:
          type: string
        numberValue:
          type: integer
        textValue:
          type: string
        timeValue:
          type: string
        type:
          type: string
...
    repo.EntityOut:
      type: object
      properties:
        archived:
          type: boolean
        assetId:
          type: string
          example: "0"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/repo.ItemAttachment"
        children:
          description: Container-specific fields (for entities whose
            entity_type.is_location = true)
          type: array
          items:
            $ref: "#/components/schemas/repo.EntitySummary"
        createdAt:
          type: string
        description:
          type: string
        entityType:
          allOf:
            - $ref: "#/components/schemas/repo.EntityTypeSummary"
          x-omitempty: true
          nullable: true
        fields:
          type: array
          items:
            $ref: "#/components/schemas/repo.EntityFieldData"
        icon:
          type: string
        id:
          type: string
        imageId:
          type: string
          x-omitempty: true
          nullable: true
        insured:
          type: boolean
        itemCount:
          description: Container-specific (populated when querying locations)
          type: number
        lifetimeWarranty:
          description: Warranty
          type: boolean
        location:
          description: |-
            Location is the nearest ancestor whose entity type is a location.
            It can differ from Parent when an item is nested in other entities.
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        manufacturer:
          type: string
        modelNumber:
          type: string
        name:
          type: string
        notes:
          description: Extras
          type: string
        parent:
          description: Edges
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        purchaseDate:
          description: Purchase
          type: string
        purchaseFrom:
          type: string
        purchasePrice:
          type: number
        quantity:
          type: number
        serialNumber:
          type: string
        soldDate:
          description: Sold
          type: string
        soldNotes:
          type: string
        soldPrice:
          type: number
        soldTo:
          type: string
        syncChildEntityLocations:
          type: boolean
        tags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TagSummary"
        thumbnailId:
          type: string
          x-omitempty: true
          nullable: true
        totalPrice:
          type: number
        updatedAt:
          type: string
        warrantyDetails:
          type: string
        warrantyExpires:
          type: string
...
    repo.EntitySummary:
      type: object
      properties:
        archived:
          type: boolean
        assetId:
          type: string
          example: "0"
        createdAt:
          type: string
        description:
          type: string
        entityType:
          allOf:
            - $ref: "#/components/schemas/repo.EntityTypeSummary"
          x-omitempty: true
          nullable: true
        icon:
          type: string
        id:
          type: string
        imageId:
          type: string
          x-omitempty: true
          nullable: true
        insured:
          type: boolean
        itemCount:
          description: Container-specific (populated when querying locations)
          type: number
        name:
          type: string
        parent:
          description: Edges
          allOf:
            - $ref: "#/components/schemas/repo.EntitySummary"
          x-omitempty: true
          nullable: true
        purchasePrice:
          type: number
        quantity:
          type: number
        soldDate:
          description: Sale details
          type: string
        tags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TagSummary"
        thumbnailId:
          type: string
          x-omitempty: true
          nullable: true
        updatedAt:
          type: string
    repo.EntityTemplateCreate:
      type: object
      required:
        - name
      properties:
        defaultDescription:
          type: string
          maxLength: 1000
          nullable: true
        defaultInsured:
          type: boolean
        defaultLifetimeWarranty:
          type: boolean
        defaultLocationId:
          description: Default location and tags
          type: string
          nullable: true
        defaultManufacturer:
          type: string
          maxLength: 255
          nullable: true
        defaultModelNumber:
          type: string
          maxLength: 255
          nullable: true
        defaultName:
          type: string
          maxLength: 255
          nullable: true
        defaultQuantity:
          description: Default values for entities
          type: number
          nullable: true
        defaultTagIds:
          type: array
          items:
            type: string
          nullable: true
        defaultWarrantyDetails:
          type: string
          maxLength: 1000
          nullable: true
        description:
          type: string
          maxLength: 1000
        fields:
          description: Custom fields
          type: array
          items:
            $ref: "#/components/schemas/repo.TemplateField"
        includePurchaseFields:
          type: boolean
        includeSoldFields:
          type: boolean
        includeWarrantyFields:
          description: Metadata flags
          type: boolean
        name:
          type: string
          maxLength: 255
          minLength: 1
        notes:
          type: string
          maxLength: 1000
    repo.EntityTemplateOut:
      type: object
      properties:
        createdAt:
          type: string
        defaultDescription:
          type: string
        defaultInsured:
          type: boolean
        defaultLifetimeWarranty:
          type: boolean
        defaultLocation:
          description: Default location and tags
          allOf:
            - $ref: "#/components/schemas/repo.TemplateLocationSummary"
        defaultManufacturer:
          type: string
        defaultModelNumber:
          type: string
        defaultName:
          type: string
        defaultQuantity:
          description: Default values for entities
          type: number
        defaultTags:
          type: array
          items:
            $ref: "#/components/schemas/repo.TemplateTagSummary"
        defaultWarrantyDetails:
          type: string
        description:
          type: string
        fields:
          description: Custom fields
          type: array
          items:
            $ref: "#/components/schemas/repo.TemplateField"
        id:
          type: string
        includePurchaseFields:
          type: boolean
        includeSoldFields:
          type: boolean
        includeWarrantyFields:
          description: Metadata flags
          type: boolean
        name:
          type: string
        notes:
          type: string
        photoMimeType:
          type: string
        photoPath:
          description: Template photo (copied as the primary photo to entities created
            from this template)
          type: string
        updatedAt:
          type: string
    repo.EntityTemplateSummary:
      type: object
      properties:
        createdAt:
          type: string
        description:
          type: string
        id:
          type: string
        name:
          type: string
        updatedAt:
          type: string
    repo.EntityTemplateUpdate:
      type: object
      required:
        - name
      properties:
        defaultDescription:
          type: string
          maxLength: 1000
          nullable: true
        defaultInsured:
          type: boolean
        defaultLifetimeWarranty:
          type: boolean
        defaultLocationId:
          description: Default location and tags
          type: string
          nullable: true
        defaultManufacturer:
          type: string
          maxLength: 255
          nullable: true
        defaultModelNumber:
          type: string
          maxLength: 255
          nullable: true
        defaultName:
          type: string
          maxLength: 255
          nullable: true
        defaultQuantity:
          description: Default values for entities
          type: number
          nullable: true
        defaultTagIds:
          type: array
          items:
            type: string
          nullable: true
        defaultWarrantyDetails:
          type: string
          maxLength: 1000
          nullable: true
        description:
          type: string
          maxLength: 1000
        fields:
          description: Custom fields
          type: array
          items:
            $ref: "#/components/schemas/repo.TemplateField"
        id:
          type: string
        includePurchaseFields:
          type: boolean
        includeSoldFields:
          type: boolean
        includeWarrantyFields:
          description: Metadata flags
          type: boolean
        name:
          type: string
          maxLength: 255
          minLength: 1
        notes:
          type: string
          maxLength: 1000
...
    repo.EntityTypeSummary:
      type: object
      properties:
        createdAt:
          type: string
        defaultTemplate:
          $ref: "#/components/schemas/repo.EntityTemplateSummary"
        defaultTemplateId:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        isContainer:
          type: boolean
        isLocation:
          type: boolean
        name:
          type: string
        updatedAt:
          type: string
...
    repo.ItemAttachment:
      type: object
      properties:
        createdAt:
          type: string
        id:
          type: string
        mimeType:
          type: string
        path:
          type: string
        primary:
          type: boolean
        thumbnail:
          $ref: "#/components/schemas/ent.Attachment"
        title:
          type: string
        type:
          type: string
        updatedAt:
          type: string
...
    repo.TagSummary:
      type: object
      properties:
        color:
          type: string
        createdAt:
          type: string
        description:
          type: string
        icon:
          type: string
        id:
          type: string
        name:
          type: string
        parentId:
          type: string
          nullable: true
        updatedAt:
          type: string
...
    repo.TemplateField:
      type: object
      properties:
        booleanValue:
          type: boolean
        id:
          type: string
        name:
          type: string
        numberValue:
          type: integer
        textValue:
          type: string
        timeValue:
          type: string
        type:
          type: string
    repo.TemplateLocationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    repo.TemplateTagSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    v1.EntityTemplateBatchCreateRequest:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          maximum: 100
          minimum: 1
        entityTypeId:
          type: string
        namePrefix:
          type: string
          maxLength: 240
        parentId:
          type: string
        startNumber:
          type: integer
          minimum: 1
        tagIds:
          type: array
          items:
            type: string
    v1.EntityTemplateCreateItemRequest:
      type: object
      required:
        - name
        - parentId
      properties:
        description:
          type: string
          maxLength: 1000
        entityTypeId:
          description: >-
            EntityTypeID is the entity type selected by the user. When set it
            takes

            precedence; when empty the repository falls back to the group's default.
          type: string
        name:
          type: string
          maxLength: 255
          minLength: 1
        parentId:
          type: string
        quantity:
          type: number
        tagIds:
          type: array
          items:
            type: string
...
    validate.ErrorResponse:
      type: object
      properties:
        error:
          type: string
        fields:
          type: string